refactor: changing visual flow for entry workers.
we no longer use an edge from a task to a worker node to indicate an entry point into the graph. instead, we are carrying a specific entryNode variable in the node definition.

this will be set true automatically when connected to a datainput node. otherwise, it can be toggled via the context menu.

to make it easier to detect entry nodes, their handle has been changed into a triangle.